HD background now fades in over low-res one
I'm using a blurry, quick-loading version of the background image to
start, while also loading the original HD-version of the image in a
totally transparent image tag. Once the image has loaded, a callback
function is triggered to fade the image in over the blurry one with CSS
transitions.

Pattern for this inspired from this: https://github.com/BojanGvozderac/react-iron-image